+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 1.0.0 (2025-08-06)
 
+* feat(hooks): 添加 useMouse Hook ([734e736](https://github.com/zjydipingxian/hooks-encode/commit/734e736))
+* feat(hooks): 添加 useWinResize Hook ([70b6cda](https://github.com/zjydipingxian/hooks-encode/commit/70b6cda))
+* feat(useDebounceFn): 增加取消当前防抖功能并优化 API ([ddc2705](https://github.com/zjydipingxian/hooks-encode/commit/ddc2705))
 * chore: publish packages ([02f9bc2](https://github.com/zjydipingxian/hooks-encode/commit/02f9bc2))
 * chore: publish packages ([52f94f8](https://github.com/zjydipingxian/hooks-encode/commit/52f94f8))
-* feat(hooks): 添加 useMouse Hook ([734e736](https://github.com/zjydipingxian/hooks-encode/commit/734e736))
-* feat(useDebounceFn): 增加取消当前防抖功能并优化 API ([ddc2705](https://github.com/zjydipingxian/hooks-encode/commit/ddc2705))